@@ -19,15 +19,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.3",
-    "@angular/common": "~13.1.3",
-    "@angular/compiler": "~13.1.3",
-    "@angular/core": "~13.1.3",
-    "@angular/forms": "~13.1.3",
-    "@angular/localize": "~13.1.3",
-    "@angular/platform-browser": "~13.1.3",
-    "@angular/platform-browser-dynamic": "~13.1.3",
-    "@angular/router": "~13.1.3",
+    "@angular/animations": "^14.0.0",
+    "@angular/common": "^14.0.0",
+    "@angular/compiler": "^14.0.0",
+    "@angular/core": "^14.0.0",
+    "@angular/forms": "^14.0.0",
+    "@angular/localize": "^14.0.0",
+    "@angular/platform-browser": "^14.0.0",
+    "@angular/platform-browser-dynamic": "^14.0.0",
+    "@angular/router": "^14.0.0",
     "bootstrap": "^5.1.3",
     "core-js": "^3.19.1",
     "@popperjs/core": "^2.9.2",
@@ -36,10 +36,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.3",
-    "@angular/cli": "~13.1.3",
-    "@angular/compiler-cli": "~13.1.3",
-    "@angular/language-service": "~13.1.3",
+    "@angular-devkit/build-angular": "^14.0.0",
+    "@angular/cli": "^14.0.0",
+    "@angular/compiler-cli": "^14.0.0",
+    "@angular/language-service": "^14.0.0",
     "@types/jasmine": "~3.10.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^14.0.0",
@@ -56,7 +56,7 @@
     "npm-run-all": "^4.1.5",
     "ts-node": "~10.4.0",
     "rimraf": "^3.0.2",
-    "typescript": "~4.4.4",
+    "typescript": "~4.7.3",
     "webpack-bundle-analyzer": "^4.5.0",
     "eslint": "^7.26.0",
     "eslint-plugin-import": "^2.22.1",
