@@ -18,7 +18,7 @@ import { ActorsService } from '../services/actors.service';
 import { MoviesService } from '../services/movies.service';
 import { UsersService } from '../services/users.service';
 import { iif, of, Observable } from 'rxjs';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { tap, debounceTime, distinctUntilChanged, switchMap, filter, catchError } from 'rxjs/operators';
 import { ActivatedRoute, Router } from '@angular/router';
 
@@ -32,12 +32,12 @@ export class SearchComponent implements OnInit, AfterViewInit {
 
     @ViewChild('movies') moviesRef: ElementRef;
     generes: Genere[];
-    movieTitle = new FormControl('');
+    movieTitle = new UntypedFormControl('');
     movies: Observable<Movie[]>;
-    movieActor = new FormControl('');
+    movieActor = new UntypedFormControl('');
     actors: Observable<Actor[]>;
     importMovies: Movie[] = [];
-    importMovieTitle = new FormControl();
+    importMovieTitle = new UntypedFormControl();
     actorsLoading = false;
     moviesLoading = false;
     importMoviesLoading = false;
